Update xivo_install_current.sh
Add a "-a" option (a stand for archive). Because sometimes I have uncomptabile VM infra with the ISO from mirror.xivo.io, so i had to use the  xivo_install_current.sh. But some other times, i want to choose my version (beacuse of compatibility with third party ecosystem: bluemind, xuc, novatax and so on...).